Rename docker.py to docker_helpers.py
The name 'docker' conflicts with Docker's python API. This is a
quick hack to unblock converting from using the Docker CLI to the
Docker API to pull and save images.

- Renamed file
- Renamed import statement from docker to docker_helpers and
gave it an alias 'dhelper'
- Replaced all instances of 'docker' with 'dhelper'
- Replaced the reference to 'import docker' with an import
from a local module in the contributing style guide

Fixes #144
